TOOL: renames file from checkbox-tree.ts to mock-data.ts
