+++ libs/core/ui/src/lib/checkbox-tree/mock-data.ts
